+++ tests/components/modal-prompt-value-with-handler.vue
@@ -6,27 +6,30 @@
 </template>
 
 <script setup lang="ts">
-import {onMounted} from "vue";
 import Modal from "../../src/utils/Modal";
+import onBeforeModalClose from "../../src/hooks/onBeforeModalClose";
 
 interface IProps {
 	value: any,
-	timeout?: number
 }
-
-const props = withDefaults(defineProps<IProps>(), {
-	timeout: 1 * 1000
-})
+const props = defineProps<IProps>()
 const emits = defineEmits<{
 	(event: typeof Modal.EVENT_PROMPT, value: any): void
 }>()
+
 function handleClick() {
 	emits(Modal.EVENT_PROMPT, props.value);
-
 }
 
-onMounted(() => {
-	setTimeout(handleClick, props.timeout)
+let canClosed = false;
+onBeforeModalClose(e => {
+	if (!canClosed) {
+		canClosed = true;
+		return false;
+	}
+	return true;
 })
+
+
 
 </script>